Updated all commands to use the pattern lecli <action> <resource>
