+++ tests/test_usage_api.py
@@ -3,15 +3,15 @@ import json
 import httpretty
 from mock import patch
 
-from lecli import usage_api
 from examples import misc_examples as misc_ex
 from examples import response_examples as resp_ex
+from lecli.usage import usage_api
 
 
 @httpretty.activate
 @patch('lecli.api_utils.get_account_resource_id')
 @patch('lecli.api_utils.get_rw_apikey')
-@patch('lecli.usage_api._url')
+@patch('lecli.usage.usage_api._url')
 def test_get_usage(mocked_url, mocked_rw_apikey, mocked_account_resource_id, capsys):
     mocked_url.return_value = misc_ex.MOCK_USAGE_URL
     mocked_rw_apikey.return_value = misc_ex.TEST_APIKEY_WITH_VALID_LENGTH
